[csc] Create registry code package
Problem:
The reconciliation of an OperatorSource no longer needs to create a
CatalogSourceConfig. With this change, the OperatorSource reconciler
will need access to the registry code that is embedded in the
catalogsourceconfig package.

Solution:
Move registry code to its own package. Change the method signatures of
these registry methods that assume the creation of an in-memory
CatalogSourceConfig (since this will no longer be the case in the
OperatorSource reconciler).