[opsrc] Combine Downloading and Configuring Phases
- An issue was found where non-idempotent phases can be scheduled to
multiple instances of the running operator. As such there were
inconsistencies between the packages listed in the operatorsource status
and the catalogsourceconfig spec because the cache of packages
used to inform the configuring phase is not shared between multiple
instances of the operator.

- To solve this issue, we have combined the Downloading and Configuring
reconcilers of the operatorsource handler into a single reconciler. This
commit also updates the unit tests for this new configuring reconciler.