Refactor to use interface client
In order to remove the CatalogSourceConfig creation from the
OperatorSource Configuring Reconciler, we will need to share code
between both configuring reconcilers. As such we need to ensure that
both reconcilers use identical types.

This commit modifies the CatalogSourceConfig Configuring Reconciler to
make use of the interface client. It also propagates the use of the
interface client to dependent methods such as the Purging Reconciler
and the registry interface.